+++ src/MongoDBServiceProvider.php
@@ -21,10 +21,9 @@ use MongoDB\Laravel\Eloquent\Model;
 use MongoDB\Laravel\Queue\MongoConnector;
 use RuntimeException;
 
-use function array_keys;
 use function assert;
 use function class_exists;
-use function implode;
+use function get_debug_type;
 use function is_string;
 use function sprintf;
 
@@ -93,38 +92,30 @@ class MongoDBServiceProvider extends ServiceProvider
                     throw new RuntimeException('GridFS adapter for Flysystem is missing. Try running "composer require league/flysystem-gridfs"');
                 }
 
-                // Reuse an existing database connection
-                if (isset($config['connection'])) {
-                    if (isset($config['mongodb_uri'])) {
-                        throw new InvalidArgumentException('In GridFS configuration, "connection" and "mongodb_uri" options cannot be set together.');
-                    }
+                $bucket = $config['bucket'] ?? null;
 
+                // Get the bucket from a factory function
+                if ($bucket instanceof Closure) {
+                    $bucket = $bucket($app, $config);
+                }
+
+                // Get the bucket from a service
+                if (is_string($bucket) && $app->has($bucket)) {
+                    $bucket = $app->get($bucket);
+                }
+
+                // Get the bucket from the database connection
+                if (is_string($bucket) || $bucket === null) {
                     $connection = $app['db']->connection($config['connection']);
                     if (! $connection instanceof Connection) {
-                        throw new InvalidArgumentException(sprintf('The database connection "%s" does not use the "mongodb" driver.', $connection['connection']));
+                        throw new InvalidArgumentException(sprintf('The database connection "%s" does not use the "mongodb" driver.', $config['connection'] ?? $app['config']['database.default']));
                     }
 
                     $bucket = $connection->getMongoClient()
                         ->selectDatabase($config['database'] ?? $connection->getDatabaseName())
                         ->selectGridFSBucket(['bucketName' => $config['bucket'] ?? 'fs', 'disableMD5' => true]);
-
-                    // Allows setting the bucket instance directly
-                } elseif (isset($config['bucket'])) {
-                    $bucket = $config['bucket'];
-                    // Resolves the "bucket" service
-                    if (is_string($bucket)) {
-                        $bucket = $app->get($bucket);
-                    } elseif ($bucket instanceof Closure) {
-                        $bucket = $bucket($app, $config);
-                    }
-
-                    if (! $bucket instanceof Bucket) {
-                        throw new InvalidArgumentException(sprintf('Provided GridFS bucket is not a instance of "%s"', Bucket::class));
-                    }
-                }
-
-                if (! isset($bucket)) {
-                    throw new InvalidArgumentException(sprintf('The "gridfs" configuration requires the "connection", "mongodb_uri", or "bucket" option. Got "%s"', implode('", "', array_keys($config))));
+                } elseif (! $bucket instanceof Bucket) {
+                    throw new InvalidArgumentException(sprintf('Unexpected value for GridFS "bucket" configuration. Expecting "%s". Got "%s"', Bucket::class, get_debug_type($bucket)));
                 }
 
                 $adapter = new GridFSAdapter($bucket, $config['prefix'] ?? '');
@@ -137,6 +128,9 @@ class MongoDBServiceProvider extends ServiceProvider
 
                     $adapter = new ReadOnlyFilesystemAdapter($adapter);
                 }
+
+                /** Prevent using backslash on Windows in {@see FilesystemAdapter::__construct()} */
+                $config['directory_separator'] = '/';
 
                 return new FilesystemAdapter(new Filesystem($adapter, $config), $adapter, $config);
             });